refactor/edit-input-payrollAPI-and-get-list-salary-by-month

+++ Validations/FindStaffAndOTRequest.cs
@@ -3,7 +3,7 @@ namespace PayrollAPI.Validations
 {
     public class FindStaffAndOTRequest
     {
-        [Required] public long staffId { get; set; }
+        [Required] public int staffId { get; set; }
         [MaxLength(10)]
         [Required] public string month { get; set; }
     }
